+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -8,7 +8,7 @@ import ru.yandex.practicum.filmorate.model.RatingMpa;
 import java.util.Collection;
 
 public interface FilmService {
-    Film getFilmById(Integer id);
+    Film getFilmById(Long id);
 
     Collection<Film> getAllFilms();
 
@@ -26,11 +26,11 @@ public interface FilmService {
 
     Film updateFilm(Film film);
 
-    Film putLike(Integer id, Long userId);
+    Film putLike(Long id, Long userId);
 
-    Film deleteLike(Integer id, Long userId);
+    Film deleteLike(Long id, Long userId);
 
-    String deleteFilm(Integer id);
+    String deleteFilm(Long id);
 
     Collection<Film> getCommonFilms(Long userId, Long friendId);
 
@@ -46,7 +46,7 @@ public interface FilmService {
 
     String deleteDirector(Integer id);
 
-    void isExistsIdFilm(Integer filmId);
+    void isExistsIdFilm(Long filmId);
 
     Collection<Film> searchFilms(String query, String by);
 }
